requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
pydantic==2.9.2
requests==2.32.3
python-dotenv==1.0.1
azure-communication-email==1.0.0
slack-sdk==3.33.4
git+https://github.com/rahul-08-11/pyzohocrm.git